dedup address request for consistency
